This introduces an alternative hardware API called nativeio structured around different functions that are typically accelerated by native hardware. Its not meant to reflect the structure of the hardware.
Docs are here: http://tannewt-micropython.readthedocs.io/en/microcontroller/

It differs from upstream's machine in the following ways:

* Python API is identical across ports due to code structure. (Lives in shared-bindings)
* Focuses on abstracting common functionality (AnalogIn) and not representing structure (ADC).
* Documentation lives with code making it easy to ensure they match.
* Pin is split into references (board.D13 and microcontroller.pin.PA17) and functionality (DigitalInOut).
* All nativeio classes claim underlying hardware resources when inited on construction, support Context Managers (aka with statements) and have deinit methods which release the claimed hardware.
* All constructors take pin references rather than peripheral ids. Its up to the implementation to find hardware or throw and exception.

+++ atmel-samd/mphalport.h
@@ -32,7 +32,7 @@
 #define USB_RX_BUF_SIZE 128
 
 // Global millisecond tick count (driven by SysTick interrupt).
-extern volatile uint32_t ticks_ms;
+extern volatile uint64_t ticks_ms;
 
 static inline mp_uint_t mp_hal_ticks_ms(void) {
   return ticks_ms;
